move input to separate class to fix type in test, update snapshots
